pyproject.toml
[project]
name = "graph-retrieval-pipeline"
version = "1.0.0"
description = "End-to-End Graph Retrieval Pipeline for Specialized Domains"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "faiss-cpu>=1.13.0",
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.7",
    "networkx>=3.6",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pydantic>=2.0.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "requests>=2.31.0",
    "scikit-learn>=1.5.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "together>=1.5.31",
    "tqdm>=4.66.0",
    "transformers>=4.40.0",
]
